Refactor app extracting helpters to a new module
