+++ app.rb
@@ -6,8 +6,9 @@ require './label'
 require './book'
 require './genre'
 require './music_album'
-
+require './helpers'
 class App
+  include Helpers
   def initialize
     @labels = []
     @genres = []
@@ -119,71 +120,6 @@ class App
     genre && game.add_genre(genre)
     puts 'Game added successfully.'
     puts ''
-  end
-
-  def ask_details(item_type)
-    print "Does your #{item_type} have an author? [y/n]: "
-    has_author = gets.chomp.downcase == 'y'
-    print "Does your #{item_type} have a label? [y/n]: "
-    has_label = gets.chomp.downcase == 'y'
-    print "Does your #{item_type} have a genre? [y/n]: "
-    has_genre = gets.chomp.downcase == 'y'
-    author = has_author ? ask_author_details : nil
-    label = has_label ? ask_label_details : nil
-    genre = has_genre ? ask_genre_details : nil
-    [author, label, genre]
-  end
-
-  def ask_author_details
-    list_authors
-    print 'Does the author appear in the list above? [y/n]: '
-    if gets.chomp.downcase == 'y'
-      print 'Enter the author ID: '
-      author_id = gets.chomp.to_i
-      return @authors.find(proc {nil}) { |author| author.id == author_id }
-    else
-      print 'Enter the author first name: '
-      first_name = gets.chomp
-      print 'Enter the author last name: '
-      last_name = gets.chomp
-      author = Author.new(first_name, last_name)
-      @authors << author
-      return author
-    end
-  end
-
-  def ask_label_details
-    list_labels
-    print 'Does the label appear in the list above? [y/n]: '
-    if gets.chomp.downcase == 'y'
-      print 'Enter the label ID: '
-      label_id = gets.chomp.to_i
-      return @labels.find(poc {nil}) { |label| label.id == label_id }
-    else
-      print 'Enter the label title: '
-      title = gets.chomp
-      print 'Enter the label color: '
-      color = gets.chomp
-      label = Label.new(title, color)
-      @labels << label
-      return label
-    end
-  end
-
-  def ask_genre_details
-    list_all_genres
-    print 'Does the genre appear in the list above? [y/n]: '
-    if gets.chomp.downcase == 'y'
-      print 'Enter the genre ID: '
-      genre_id = gets.chomp.to_i
-      return @genres.find { |genre| genre.id == genre_id }
-    else
-      print 'Enter the genre name: '
-      name = gets.chomp
-      genre = Genre.new(name)
-      @genres << genre
-      return genre
-    end
   end
 
   def load_resource(resource)
